+++ db/upgrade/2.5.1-upgrade-2.6.0-mysql.sql
@@ -168,3 +168,13 @@ INSERT INTO `plugin_handle` VALUES ('1678996921914392576', '42', 'loadBalance', 
 INSERT INTO `plugin_handle` VALUES ('1678997769998467072', '42', 'clientMaxLifeTimeMs', 'clientMaxLifeTimeMs', 2, 1, 8, '{\"required\":\"0\",\"defaultValue\":\"60000\",\"rule\":\"\"}', '2023-01-10 10:08:01.158', '2023-01-10 10:08:01.158');
 INSERT INTO `plugin_handle` VALUES ('1678997277012557824', '42', 'clientMaxConnections', 'clientMaxConnections', 2, 1, 6, '{\"required\":\"0\",\"defaultValue\":\"20\",\"rule\":\"\"}', '2023-01-10 10:08:01.158', '2023-01-10 10:08:01.158');
 INSERT INTO `plugin_handle` VALUES ('1678997557628272640', '42', 'clientPendingAcquireTimeout', 'clientPendingAcquireTimeout', 2, 1, 5, '{\"required\":\"0\",\"defaultValue\":\"5\",\"rule\":\"\"}', '2023-01-10 10:08:01.158', '2023-01-10 10:08:01.158');
+
+/* motan rpc */
+UPDATE `plugin` set config = '{\"registerProtocol\":\"zk\",\"registerAddress\":\"127.0.0.1:8002\",\"corethreads\":0,\"threads\":2147483647,\"queues\":0,\"threadpool\":\"shared\"}' WHERE id = '17';
+UPDATE `plugin_handle` set field = 'registerProtocol', label = 'registerProtocol', sort = 0, ext_obj = '{\"required\":\"0\",\"defaultValue\":\"direct\",\"placeholder\":\"registerProtocol\",\"rule\":\"\"}' WHERE id = '1529402613204172834';
+UPDATE `plugin_handle` set sort = 2 where id = '1529402613204172835';
+UPDATE `plugin_handle` set sort = 3 where id = '1529402613204172836';
+UPDATE `plugin_handle` set sort = 4 where id = '1529402613204172837';
+UPDATE `plugin_handle` set sort = 5 where id = '1529402613204172838';
+
+INSERT INTO `plugin_handle` VALUES ('1678997557628272641', '17', 'registerAddress', 'registerAddress', 2, 3, 1, '{\"required\":\"0\",\"defaultValue\":\"127.0.0.1:8002\",\"placeholder\":\"registerAddress\",\"rule\":\"\"}', '2023-01-10 10:08:01.158', '2023-01-10 10:08:01.158'));