+++ shenyu-client/shenyu-client-dubbo/shenyu-client-apache-dubbo/src/main/java/org/apache/shenyu/client/apache/dubbo/ApacheDubboApiBeansExtractor.java
@@ -21,7 +21,6 @@ import org.apache.dubbo.config.annotation.DubboService;
 import org.apache.dubbo.config.annotation.Service;
 import org.apache.shenyu.client.apache.dubbo.processor.extractor.DubboServiceProcessor;
 import org.apache.shenyu.client.apache.dubbo.processor.extractor.ServiceProcessor;
-import org.apache.shenyu.client.apache.dubbo.processor.register.ShenyuDubboProcessor;
 import org.apache.shenyu.client.core.register.extractor.BaseAnnotationApiBeansExtractor;
 import org.apache.shenyu.client.core.register.extractor.RpcApiBeansExtractor;
 import org.apache.shenyu.common.enums.RpcTypeEnum;
@@ -37,7 +36,6 @@ public class ApacheDubboApiBeansExtractor extends BaseAnnotationApiBeansExtracto
         
         addExtractorProcessor(new DubboServiceProcessor());
         addExtractorProcessor(new ServiceProcessor());
-        addExtractorProcessor(new ShenyuDubboProcessor());
     }
     
     @Override