+++ RELEASE-NOTES.md
@@ -19,7 +19,7 @@
 15. Add naocs discovery sync
 16. Add apollo discovery sync
 17. Add HttpLongPolling discovery sync
-18. Add consul discove sync
+18. Add consul discovery sync
 19. Add huawei cloud lts loging plugin
 20. Support openGauss database for shenyu admin
 21. Support polaris config sync and register center
@@ -64,7 +64,13 @@
 34. Update apollo client register repository active offline
 35. Storage adjustment for swagger type API documents, change from local cache to database
 36. Support nacos client offline
-37. Add e2e alibaba dubbo case
+37. Add e2e alibaba dubbo test case
+38. Add e2e apache dubbo test case
+39. Add shenyu-spring-sdk test cases
+40. Add e2e sofa test cases
+41. Add apollo config sync test case
+42. Add database connection pool
+43. Add idea icon for shenyu project
 
 ### Refactor
 
@@ -119,7 +125,9 @@
 49. Refactor shenyu log observability
 50. Add bootstrap start log
 51. Refactor api document for swagger
-52. Upgrade grpc version to 1.48.0
+52. Upgrade grpc version to 1.53.0
+53. Refactor api meta data process function
+54. polish master code and pom
 
 ### Bug Fix
 
@@ -175,6 +183,9 @@
 50. Fix http registration metadata
 51. Fix websocket lost the user-defined clost status
 52. Fix consul register lose the prop of meta-path when special symbol
+53. Fix etcd sync error
+54. Fix admin sync error
+55. Fix shenyu motan plugin execute error
 
 ## 2.5.1